feat: refect javascript for web view
